refactor(isEmpty): move isEmpty under KitchenSink
